Buff Special potion, added small bit of lore, work on the boss a bit, add more enchantment and fix some enchantment, Heart of Earth fix, added new debug item, work on post boss fight and ect

+++ Contents/Artifacts/HeartOfEarthArtifact.cs
@@ -23,7 +23,7 @@ namespace BossRush.Contents.Artifacts {
 		}
 		public override void ModifyWeaponDamage(Item item, ref StatModifier damage) {
 			if (Earth) {
-				damage *= Player.statLife / Player.statLifeMax2;
+				damage *= Player.statLife / (float)Player.statLifeMax2;
 			}
 		}
 		public override void PostUpdate() {
